// frontend/src/api/generated/model/guestDocumentType.ts
export type GuestDocumentType = typeof GuestDocumentType[keyof typeof GuestDocumentType];


export const GuestDocumentType = {
  internal_passport: 'internal_passport',
  foreign_passport: 'foreign_passport',
  residence_permit: 'residence_permit',
  student_card: 'student_card',
  driving_licence: 'driving_licence',
} as const;