// frontend/src/api/generated/model/guestVisitStatus.ts
export type GuestVisitStatus = typeof GuestVisitStatus[keyof typeof GuestVisitStatus];


export const GuestVisitStatus = {
  in_building: 'in_building',
  closed: 'closed',
  overdue: 'overdue',
  closed_late: 'closed_late',
} as const;